Cleaning up various type issues, thanks to Pan Deng and Yixiong Zou.
CVS patchset: 2351
CVS date: 2004/01/29 11:13:30


Mercurial revision: 8a843ad240f01afb18c3f2ef228c7957402fe46c

+++ telecom/checkpointd/message.c
@@ -1465,7 +1465,7 @@ SaCkptMessage2Hamsg(SaCkptMessageT* ckptMsg) {
 	
 	if (ckptMsg->paramLength > 0) {
 		strTemp[0] = 0;
-		sprintf(strTemp, "%d", ckptMsg->paramLength);
+		sprintf(strTemp, "%d", (int)ckptMsg->paramLength);
 		rc = ha_msg_mod(haMsg, F_CKPT_PARAM_LENGTH, strTemp);
 		if (rc != HA_OK) {
 			cl_log(LOG_ERR, 
@@ -1486,7 +1486,7 @@ SaCkptMessage2Hamsg(SaCkptMessageT* ckptMsg) {
 	
 	if (ckptMsg->dataLength > 0) {
 		strTemp[0] = 0;
-		sprintf(strTemp, "%d", ckptMsg->dataLength);
+		sprintf(strTemp, "%d", (int)ckptMsg->dataLength);
 		rc = ha_msg_mod(haMsg, F_CKPT_DATA_LENGTH, strTemp);
 		if (rc != HA_OK) {
 			cl_log(LOG_ERR,
